// frontend/src/lib/types.ts
export interface CaseType {
  id: string;
  title: string;
  description: string;
  status: string;
  win: boolean;
  client: {
    id: string;
    firstName: string;
    lastName: string;
    email: string;
    phone_number: string;
    address: string;
  };
  Hearings: {
    id?: string;
    createdAt: Date;
    updatedAt: Date;
    status: string;
  }[];
  Timeline: {
    id: string;
    title: string;
    description: string;
    status: string;
    createdAt: Date;
    updatedAt: Date;
  }[];
  CaseLawyer: {
    lawyer: {
      id: string;
      user: {
        firstName: string;
        lastName: string;
        email: string;
        phone_number: string;
        address: string;
      };
    };
  }[];
};

export enum CASE_STATUS {
  ACTIVE = 'ACTIVE',
  CLOSED = 'CLOSED',
  ONGOING = 'ONGOING'
}

export enum TimelineStatus {
  ACTIVE = 'ACTIVE',
  CLOSED = 'CLOSED',
  ONGOING = 'ONGOING'
}


export interface TimelineType {
  id: string;
  title: string;
  description: string;
  status: TIME_LINE_STATUS;
  createdAt: Date;
}

export interface HearingType {
  id?: string;
  client_id: string;
  date: Date;
  status: HEARING_STATUS;
  lawyer_id: string;
}

export enum HEARING_STATUS {
  UPCOMING = 'UPCOMING',
  CLOSED = 'CLOSED',
}

export enum TIME_LINE_STATUS {
  ACTIVE = 'ACTIVE',
  CLOSED = 'CLOSED',
  ONGOING = 'ONGOING',
}